Major changes. ++ Added MySQL Classes and functionalities.
It's important to clarify that the interfaces:
    - IPlayer <- FPlayer
    - IFaction <- Faction

And the classes related to those interfaces it's going to be completely
removed and replaces with the classes in the .mysql.* package.

The reason for that is that those interfaces, abstracts classes and logic
was completely out of the sight of this project.

THIS IS NOT A STABLE OF COMPILABLE VERSION OF THE PROJECT.